🩹 chore: 하단바 layout 적용
