+++ app/(route)/(bottom-nav)/layout.tsx
@@ -1,0 +1,13 @@
+import React, { ReactNode } from "react";
+import BottomNav from "@/components/BottomNav";
+
+const BottomNavLayout = ({ children }: { children: ReactNode }) => {
+  return (
+    <div>
+      <BottomNav />
+      {children}
+    </div>
+  );
+};
+
+export default BottomNavLayout;
