+++ app/(route)/(bottom-nav)/mypage/_components/tab/ArtistTab.tsx
