+++ app/(route)/(bottom-nav)/page.tsx
@@ -1,8 +1,7 @@
 import ArtistList from "@/components/artist-list/ArtistList";
-import BottomNav from "../_components/BottomNav";
-import FavArtistEventsCarousel from "./_components/carousel/FavArtistEventsCarousel";
-import NewestEventsCarousel from "./_components/carousel/NewestEventsCarousel";
-import PopularEventsCarousel from "./_components/carousel/PopularEventsCarousel";
+import FavArtistEventsCarousel from "../_components/carousel/FavArtistEventsCarousel";
+import NewestEventsCarousel from "../_components/carousel/NewestEventsCarousel";
+import PopularEventsCarousel from "../_components/carousel/PopularEventsCarousel";
 
 const Home = () => {
   return (
@@ -15,7 +14,6 @@ const Home = () => {
           <ArtistList />
         </div>
       </main>
-      <BottomNav />
     </>
   );
 };
