+++ app/(route)/(bottom-nav)/search/page.tsx
